move classes to library
