+++ image_pub_sub/src/image_pub_node.cpp
@@ -1,203 +1,14 @@
 #include <filesystem>
 #include <algorithm>
-#include <vector>
 #include <cv_bridge/cv_bridge.h>
 #include <opencv2/opencv.hpp>
 #include <ros/ros.h>
 #include <std_msgs/Header.h>
 #include <sensor_msgs/Image.h>
 #include <image_transport/image_transport.h>
+#include "image_pub_sub/factory.hpp"
 
 namespace fs = std::filesystem;
-
-/**
- * @brief A base class provides some common methods to read image from a source
- */
-class ImageReader
-{
-public:
-    /**
-     * @brief Get an image from a source
-     * @return an image if the source is readable, otherwise an std::nullopt
-     */
-    virtual std::optional<cv::Mat> get() = 0;
-
-    /**
-     * @brief Set looping so that when the source reaches its end. It can go back
-     * to the begining
-     * @param isLoop whether to loop or not
-     */
-    void setLoop(bool isLoop)
-    {
-        this->_isLoop = isLoop;
-    }
-
-    virtual size_t size() = 0;
-
-    bool isLoop()
-    {
-        return this->_isLoop;
-    }
-
-private:
-    bool _isLoop = false;
-};
-
-/**
- * @brief A class to read images from a folder containing ONLY images.
- * Since we only list all the file in the given directory, it might not
- * work if this folder contains a strange file, i.e. a text file.
- */
-class FolderImageReader
-    : public ImageReader
-{
-public:
-    /**
-     * @brief A default constructor
-     * @param imagePath a directory path
-     * @param isDisableSort whether to sort the paths or not. Use with your caution.
-     */
-    FolderImageReader(fs::path imagePath, bool isDisableSort = false)
-        : _curIndex{0}
-    {
-        if (!fs::is_directory(imagePath))
-            throw std::runtime_error("imagepath is not a directory");
-
-        auto dirIter = fs::directory_iterator(imagePath);
-        for (const auto &path : dirIter)
-        {
-            if (path.is_regular_file())
-            {
-                _imagePaths.push_back(path.path());
-            }
-        }
-
-        if (_imagePaths.empty())
-            throw std::runtime_error("image folder is empty");
-
-        if (!isDisableSort)
-            std::sort(_imagePaths.begin(), _imagePaths.end());
-    }
-
-    virtual std::optional<cv::Mat> get() override
-    {
-        if (_curIndex >= _imagePaths.size())
-        {
-            if (isLoop())
-            {
-                _curIndex = 0;
-            }
-            else
-            {
-                return std::nullopt;
-            }
-        }
-        const fs::path &imagePath = _imagePaths[_curIndex++];
-        ROS_DEBUG_STREAM("Reading from " << imagePath);
-        return cv::imread(imagePath);
-    }
-
-    virtual size_t size() override
-    {
-        return this->_imagePaths.size();
-    }
-
-private:
-    size_t _curIndex;
-    std::vector<fs::path> _imagePaths;
-};
-
-/**
- * @brief A class to read image from a singular file.
- * It is convenient for debugging when you just want
- * this node to publish a single image for the whole life time.
- */
-class FileImageReader
-    : public ImageReader
-{
-public:
-    /**
-     * @brief A default constructor
-     * @param imagePath a path to an image file. Only supports types that
-     * OpenCV supports.
-     */
-    FileImageReader(fs::path imagePath)
-        : _isReadOnce{false}
-    {
-        if (!fs::is_regular_file(imagePath))
-            throw std::runtime_error("image path is not a regular file");
-        _image = cv::imread(imagePath);
-        if (_image.empty())
-            throw std::runtime_error("image path is not an image");
-    }
-
-    virtual std::optional<cv::Mat> get() override
-    {
-        if (_isReadOnce)
-        {
-            if (isLoop())
-                return _image;
-            return std::nullopt;
-        }
-
-        _isReadOnce = true;
-        return _image;
-    }
-
-    virtual size_t size() override
-    {
-        return 1l;
-    }
-
-private:
-    bool _isReadOnce;
-    cv::Mat _image;
-};
-
-/**
- * @brief A default ImageReader factory
- * This class provide an easy to use that hiding if/else from
- * the main function.
- */
-class ImageReaderFactory
-{
-public:
-    /**
-     * @brief Open an ImageReader given an image path
-     * @param imagePath the image path. It could be a single image file
-     * or a directory containing images.
-     */
-    static std::shared_ptr<ImageReader> open(const fs::path &imagePath)
-    {
-        try
-        {
-            if (!fs::exists(imagePath))
-            {
-                throw std::runtime_error("image path does not exist");
-            }
-            if (fs::is_directory(imagePath))
-            {
-                bool isDisableSort = false;
-                ros::param::get("~disable_sort", isDisableSort);
-                return std::make_shared<FolderImageReader>(imagePath, isDisableSort);
-            }
-            else if (fs::is_regular_file(imagePath))
-            {
-                return std::make_shared<FileImageReader>(imagePath);
-            }
-            else
-            {
-                throw std::runtime_error("image path is not supported.");
-            }
-        }
-        catch (const std::exception &e)
-        {
-            ROS_FATAL_STREAM(e.what());
-            return nullptr;
-        }
-        return nullptr;
-    }
-};
 
 int main(int argc, char **argv)
 {
